refactor(bandingmachine): separate TouchScreen and Printer
separate BandingMachine to two part bc it was too heavy
 - change name of Bandingmachine TouchScreen
 - separate printing result part from Bandingmachine as Printer
