+++ src/main/java/lotto/MessageCenter.java
@@ -6,7 +6,9 @@ public enum MessageCenter {
     ERROR("[ERROR] "),
     ERROR_NOTNULL("[ERROR] 기추첨결과가 존재합니다."),
     ERROR_NULL("[ERROR] 추첨결과 저장 과정에서 에러가 발생하였습니다."),
-    ERROR_MONEY("[ERROR] 금액을 잘못 입력했습니다. 다시 입력해 주세요.");
+    ERROR_MONEY("[ERROR] 금액을 잘못 입력했습니다. 다시 입력해 주세요."),
+    ERROR_USERSTORAGE("[ERROR] 구매기록이 없습니다.");
+
 
     private final String msg;
